+++ src/main/java/codesquad/application/domain/User.java
@@ -1,4 +1,4 @@
-package codesquad.model;
+package codesquad.application.domain;
 
 public class User {
     private final String userId;
